% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalKmeansRP.R
\name{compute_tightness}
\alias{compute_tightness}
\title{Compute Tightness Criterion for Functional K-Means}
\usage{
compute_tightness(X_projected, cluster_labels, k_clusters)
}
\arguments{
\item{X_projected}{Matrix. n x d0 matrix of projected data}

\item{cluster_labels}{Vector. Cluster assignments (length n)}

\item{k_clusters}{Integer. Number of clusters}
}
\value{
Numeric. The tightness value
}
\description{
Computes the weighted tightness criterion from equation (4.2) in the manuscript:
\deqn{T̂(Î_1, Î_2 | Ψ) = \sum_{j=1}^{d_0} \frac{1}{\hat{\sigma}^2(\psi_j)}
      \frac{1}{n} \sum_{k=1}^K \sum_{i \in Î_k} [X_{i,j}(\psi_j) - \bar{X}_{k,j}(\psi_j)]^2}
}
\details{
The tightness criterion standardizes each dimension by its variance before
computing the within-cluster sum of squares. This prevents dimensions with
large variance from dominating the clustering.
}
\keyword{internal}
