% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{create_fourier_basis}
\alias{create_fourier_basis}
\title{Create Fourier Basis Functions}
\usage{
create_fourier_basis(r, grid)
}
\arguments{
\item{r}{Integer. Number of basis functions}

\item{grid}{Vector. Grid points for evaluation}
}
\value{
Matrix of r x length(grid) with basis functions in rows
}
\description{
Create Fourier Basis Functions
}
