% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalKmeansRP.R
\name{evaluate_clustering}
\alias{evaluate_clustering}
\title{Evaluate Clustering Accuracy}
\usage{
evaluate_clustering(predicted, true)
}
\arguments{
\item{predicted}{Vector. Predicted cluster labels}

\item{true}{Vector. True cluster labels}
}
\value{
List containing:
  \item{accuracy}{Overall accuracy (best permutation)}
  \item{confusion_matrix}{Confusion matrix}
  \item{adjusted_rand_index}{Adjusted Rand Index (if mclust available)}
}
\description{
Computes clustering accuracy metrics when true labels are known
}
