% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{get_basis_functions}
\alias{get_basis_functions}
\title{Extract Basis Functions from Random Projections Object}
\usage{
get_basis_functions(rp)
}
\arguments{
\item{rp}{Object of class random_projections from generateRP()}
}
\value{
Matrix of r x grid_length with basis functions in rows
}
\description{
Extract Basis Functions from Random Projections Object
}
\examples{
\dontrun{
  Q_B <- generateRP(B = 10, d0 = 2, r = 5)
  basis <- get_basis_functions(Q_B)
  dim(basis)  # 5 x 100
}

}
