% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{get_projection_functions}
\alias{get_projection_functions}
\title{Get Projection Functions as Matrix}
\usage{
get_projection_functions(rp, which_projection = 1)
}
\arguments{
\item{rp}{Object of class random_projections from generateRP()}

\item{which_projection}{Integer. Which projection to extract (default: 1)}
}
\value{
Matrix of d0 x grid_length with projection functions in rows
}
\description{
Get Projection Functions as Matrix
}
\examples{
\dontrun{
  Q_B <- generateRP(B = 10, d0 = 3, r = 5)
  proj_funcs <- get_projection_functions(Q_B, which_projection = 1)
  dim(proj_funcs)  # 3 x 100
}

}
