% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongiClust-package.R
\docType{package}
\name{LongiClust-package}
\alias{LongiClust}
\alias{LongiClust-package}
\title{LongiClust: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jiadong Mao \email{chiatungmao@gmail.com}

Authors:
\itemize{
  \item Saritha Kodikara
}

}
\keyword{internal}
