% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalKmeansRP.R
\name{plot.functionalKmeansRP}
\alias{plot.functionalKmeansRP}
\title{Plot Method for functionalKmeansRP}
\usage{
\method{plot}{functionalKmeansRP}(x, type = "all", ...)
}
\arguments{
\item{x}{Object of class functionalKmeansRP}

\item{type}{Character. Type of plot: "tightness", "clusters", "projection", or "all"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot Method for functionalKmeansRP
}
