% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{print.random_projections}
\alias{print.random_projections}
\title{Print method for random_projections}
\usage{
\method{print}{random_projections}(x, ...)
}
\arguments{
\item{x}{Object of class random_projections}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for random_projections
}
