% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{project_functional_data}
\alias{project_functional_data}
\title{Evaluate Projection Function on Functional Data}
\usage{
project_functional_data(X, psi)
}
\arguments{
\item{X}{Matrix. n x grid_length matrix of functional observations}

\item{psi}{List. Projection function from generate_random_projections}
}
\value{
Vector of n projected values
}
\description{
Evaluate Projection Function on Functional Data
}
