% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{project_multivariate_data}
\alias{project_multivariate_data}
\title{Project Multivariate Functional Data onto Psi}
\usage{
project_multivariate_data(X_list, Psi_star)
}
\arguments{
\item{X_list}{List of length d0. Each element is n x grid_length matrix}

\item{Psi_star}{List. One element from Q_B (contains d0 projection functions)}
}
\value{
Matrix of n x d0 projected values
}
\description{
Project Multivariate Functional Data onto Psi
}
