% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalKmeansRP.R
\name{summary.functionalKmeansRP}
\alias{summary.functionalKmeansRP}
\title{Summary Method for functionalKmeansRP}
\usage{
\method{summary}{functionalKmeansRP}(object, ...)
}
\arguments{
\item{object}{Object of class functionalKmeansRP}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for functionalKmeansRP
}
