% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{summary.random_projections}
\alias{summary.random_projections}
\title{Summary Method for Random Projections}
\usage{
\method{summary}{random_projections}(object, ...)
}
\arguments{
\item{object}{Object of class random_projections}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for Random Projections
}
